requirements.txt
Flask==3.1.1
numpy==1.24.4
pandas==2.2.3
Pillow==11.2.1
pyheif==0.8.0
torch==2.7.0
transformers==4.41.1
